third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/cssom/cssstyledeclaration-csstext-blocked.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
    <meta http-equiv="Content-Security-Policy" content="style-src 'unsafe-inline' 'self'">
</head>
<body>
    <script>
        description("Setting CSSStyleDeclaration::cssText blocked if 'unsafe-eval' is not an allowed source of style.");
        shouldThrow("document.head.style.cssText = 'display: none;';");
    </script>
</body>
</html>
