third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/cssom/cssstyledeclaration-csstext-blocked-expected.txt

CONSOLE ERROR: Refused to evaluate a string as CSS because 'unsafe-eval' is not an allowed source of style in the following Content Security Policy directive: "style-src 'unsafe-inline' 'self'".

Setting CSSStyleDeclaration::cssText blocked if 'unsafe-eval' is not an allowed source of style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.head.style.cssText = 'display: none;'; threw exception SecurityError: Failed to set the 'cssText' property on 'CSSStyleDeclaration': Refused to evaluate a string as CSS because 'unsafe-eval' is not an allowed source of style in the following Content Security Policy directive: "style-src 'unsafe-inline' 'self'".
.
PASS successfullyParsed is true

TEST COMPLETE
